Add TDD solutions
I came up with these by implementing each requirement in turn without
going through the walk through. Some of the differences are due to my
lack of familiarity with the Java standard library (I'm normally a big
fan of pipelines!), but I found the process easy to follow all the same.

FWIW, note that my implementation allows for arbitrary regexes as
delimiters, rather than replacing delimiters as commas and splitting the
string only on commas.